refactor: 체스보드 초기 기물 셋팅 로직을 Generator로 분리
