+++ src/main/java/chess/ConsoleApplication.java
@@ -7,7 +7,7 @@ import chess.domain.board.ChessBoardGenerator;
 public class ConsoleApplication {
 
     public static void main(String[] args) {
-        ChessBoard chessBoard = new ChessBoard(ChessBoardGenerator.generateEmptyBoard());
+        ChessBoard chessBoard = new ChessBoard(ChessBoardGenerator.generateDefaultChessBoard());
         ChessController chessController = new ChessController(chessBoard);
         chessController.run();
     }
